+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- Introduce `useBrowserFocus` option to use the browser's native focus management instead of the geocoder's custom focus management. This is useful for accessibility.
+
 ## 5.0.3
 
 ### Features / Improvements 🚀